\name{ShiraziBoersma-package}
\alias{ShiraziBoersma-package}
\alias{ShiraziBoersma}
\docType{package}
\title{
\packageTitle{Title of the package}
}
\description{
\packageDescription{ShiraziBoersma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShiraziBoersma}
\packageIndices{ShiraziBoersma}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ShiraziBoersma}

Maintainer: \packageMaintainer{ShiraziBoersma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
