\name{ShiraziBoersma}
\alias{ShiraziBoersma}
\docType{package}
\title{Shirazi and Boersma TT}
\description{
  Shirazi and Boersma texture triangle
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShiraziBoersma}
\packageIndices{ShiraziBoersma}
Just do it
}
\author{
\packageAuthor{Edward Jones}
}

