\name{ShiraziBoersma}
\alias{ShiraziBoersma}
\docType{package}
\title{Shirazi and Boersma TT}
\description{
  Shirazi and Boersma texture triangle
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Calculate mean geometric diameter and standard deviation based on particle 
                    size analaysis data based on Shirazi et al. (1988). Plot Australian, IUSS 
                    and USDA texture triangles as mean gemometric diamter and standard deviation.
}
}
\author{
  Edward Jones <edward.jones@sydney.edu.au>
}
